requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
bcrypt==4.3.0
bitarray==3.6.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
ckzg==2.1.1
click==8.2.1
covalent-api-sdk==1.0.2
cytoolz==1.0.1
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
fastapi==0.116.1
frozenlist==1.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
hexbytes==1.3.1
httplib2==0.22.0
idna==3.10
iniconfig==2.1.0
multidict==6.6.3
packaging==25.0
parsimonious==0.10.0
passlib==1.7.4
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-env==1.1.5
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pyunormalize==16.0.0
regex==2025.7.34
requests==2.32.4
rlp==4.1.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
toolz==1.0.0
tqdm==4.67.1
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
web3==7.13.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1
